Reorganize package structure and add gltfpack as a compression option.
