+++ glue_ar/common/common.py
