+++ glue_ar/common/scatter.py
