+++ glue_ar/common/scatter_export_options.py
@@ -1,6 +1,6 @@
 from echo import CallbackProperty
 from glue.core.state_objects import State
-from glue_ar.export_dialog import ar_layer_export
+from glue_ar.common.export_state import ar_layer_export
 from glue_vispy_viewers.scatter.layer_state import ScatterLayerState
 
 __all__ = ["ARScatterExportOptions"]
