+++ glue_ar/qt/export_dialog.ui
@@ -14,18 +14,15 @@
    <string>Export 3D Volume</string>
   </property>
   <layout class="QGridLayout" name="gridLayout">
-   <item row="7" column="0">
-    <widget class="QCheckBox" name="bool_draco">
-     <property name="text">
-      <string>Use Draco compression</string>
-     </property>
-    </widget>
-   </item>
-   <item row="6" column="0">
-    <widget class="QComboBox" name="combosel_filetype"/>
-   </item>
    <item row="3" column="0">
     <widget class="QListWidget" name="listsel_layer"/>
+   </item>
+   <item row="2" column="0">
+    <widget class="QLabel" name="label_settings_message">
+     <property name="text">
+      <string>Set the export settings for each layer</string>
+     </property>
+    </widget>
    </item>
    <item row="5" column="0">
     <widget class="QLabel" name="label_filetype_message">
@@ -34,10 +31,13 @@
      </property>
     </widget>
    </item>
-   <item row="2" column="0">
-    <widget class="QLabel" name="label_settings_message">
+   <item row="6" column="0">
+    <widget class="QComboBox" name="combosel_filetype"/>
+   </item>
+   <item row="7" column="0">
+    <widget class="QLabel" name="label_compression_message">
      <property name="text">
-      <string>Set the export settings for each layer</string>
+      <string>Select compression method</string>
      </property>
     </widget>
    </item>
@@ -82,6 +82,9 @@
    </item>
    <item row="4" column="0">
     <layout class="QFormLayout" name="layer_layout"/>
+   </item>
+   <item row="8" column="0">
+    <widget class="QComboBox" name="combosel_compression"/>
    </item>
   </layout>
  </widget>
