+++ glue_ar/qt/qr.py
