+++ glue_ar/qt/qr_dialog.ui
