--- glue_ar/tools/__init__.py
@@ -1,2 +1,0 @@
-from .jupyter import *  # noqa
-from .qt import *  # noqa
